requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
coinpayments-py==0.1.0
Django==3.1
django-coinpayments==0.1.3
django-model-utils==4.1.1
djangorestframework==3.12.4
idna==2.10
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
